requirements.txt
langchain
openai
python-dotenv
tiktoken
faiss-cpu
validators
streamlit
pypdf